Update navigation bar
Update main.py and multi_app.py
